- updated longs to decimal or int

+++ Source/Flinks.CSharp.SDK.Model/Shared/Account.cs
@@ -10,19 +10,19 @@ namespace Flinks.CSharp.SDK.Model.Shared
     public class Account
     {
         [JsonProperty("EftEligibleRatio")]
-        public long EftEligibleRatio { get; set; }
+        public decimal EftEligibleRatio { get; set; }
 
         [JsonProperty("Transactions")]
         public List<Transaction> Transactions { get; set; }
 
         [JsonProperty("TransitNumber")]
-        public long? TransitNumber { get; set; }
+        public int? TransitNumber { get; set; }
 
         [JsonProperty("InstitutionNumber")]
-        public long? InstitutionNumber { get; set; }
+        public int? InstitutionNumber { get; set; }
 
         [JsonProperty("OverdraftLimit", NullValueHandling = NullValueHandling.Ignore)]
-        public long? OverdraftLimit { get; set; }
+        public decimal? OverdraftLimit { get; set; }
 
         [JsonProperty("Title")]
         public string Title { get; set; }
